Remove IAppxFile from StreamBase. Most of the derived classed of StreamBase don't care about if it is an IAppxFile or not and we end up with a lot of them just returning E_NOTIMPL. This is also confusing sometimes a stream can QI for IAppxFile and succeed, but if we ever get an outside stream and asume that is an appx file we will have problems.

IAppxFileInternal had nothing to do with AppxFiles, so I remove it. Its only functionality was to return the size of the stream in the zip container to validate the information of the block map. IStreamInternal::GetSizeOnZip provides this functionality now.

Because the vtables are smaller, this change also reduce the size of the binary ~7k

Related work items: #18752968, #18853644

+++ src/msix/AppxPackageInfo.cpp
@@ -169,7 +169,7 @@ namespace MSIX {
         memcpy(buffer.data(), &wpublisher[0], wpublisher.size() * sizeof(char16_t));
 
         std::vector<std::uint8_t> hash;
-        ThrowErrorIfNot(Error::Unexpected, SHA256::ComputeHash(buffer.data(), buffer.size(), hash),  "Failed computing publisherId");
+        ThrowErrorIfNot(Error::Unexpected, SHA256::ComputeHash(buffer.data(), static_cast<uint32_t>(buffer.size()), hash),  "Failed computing publisherId");
         return Base32Encoding(hash);
     }
